added attribute to User to GameStat and thereby creating a direct relation between GameStat and User.
The original idea was flawed. The idea was that since a User contains game, and game contains gamestat, I could simply retrieve the gamestat through game which the user had accesses to, thereby avoiding duplicate values for userId in both game and gamestat.

But I could never retrieve the users gameStat since it had no unique identifier for a particular user. It only had information about which game it is a part of, and User only had information about which game the user was a part of. So Any user that was a part of a given game could be the owner of any gamestat. Again there was no unique identifier.